% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfa.discovery.r
\name{erfa.discovery}
\alias{erfa.discovery}
\title{Plot pathway representations obtain from ERFA model}
\usage{
erfa.discovery(
  rg,
  regulation.step,
  regulation.flow,
  gene.from,
  gene.to,
  plot.dir,
  plot.file.name.base = NULL,
  plot.title.base = NULL,
  pathway.regulation.approx = 0.05,
  pathway.n.max = 10,
  pathway.gene.max = 20,
  alpha = NULL,
  plot.moa = F
)
}
\arguments{
\item{rg}{An \link{igraph} graph object with the interaction graph, usually
obtained \code{\link[=build.regulation.graph]{build.regulation.graph()}}.}

\item{regulation.step}{A matrix with one-step regulation flow values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{regulation.flow}{A matrix with all-step regulation flow values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{gene.from}{Character vector with gene names as source of regulation.}

\item{gene.to}{Character vector with gene names as target of regulation.
Elements also in \code{gene.from} are ignored.}

\item{plot.dir}{Character string with the directory to save the plot.}

\item{plot.file.name.base}{Character string with a prefix for the plot file
names. If \code{NULL}, default values are generated depending on the number
of genes in \code{gene.from} and \code{gene.to}.}

\item{plot.title.base}{Character string with a prefix for the plot titles.
If \code{NULL}, default values are generated depending on the number of genes
in \code{gene.from} and \code{gene.to}.}

\item{pathway.regulation.approx}{Float with the desired approximation of
pathway coverage.}

\item{pathway.n.max}{Integer with the maximum number of pathways to plot.}

\item{pathway.gene.max}{Integer with the maximum number of genes to be
plotted in any pathway.}

\item{alpha}{a vector containing relative perturbation magnitudes for
source of perturbation. If NULL, alpha is set to a vector with
components are equal to 1 for gene.from
genes and 0 for other genes.}

\item{plot.moa}{Should pathways be saved as plots to PDF files?}
}
\value{
\code{NULL}, invisibly.
}
\description{
The function plots partial pathway representations between genes or set of genes, in
increasing amounts of genes included and flow explained.
}
\details{
The call is ignored if \code{gene.from} or \code{gene.to}, after removing overlapping
values from \code{gene.to}, is \code{NULL} or has no elements.
}
\examples{
import.database.omnipath("./database_omnipath.rda", TRUE)
interacs <- get.interactions( "omnipath", 
                              "./database_omnipath.rda", "./table_interacs.csv",
                              "./cache_interacs.rda", FALSE )

reg.graph <- build.regulation.graph( interacs, NULL, "./cache_reg_graph.rda", 
                                     FALSE )

reg.matrix <- calculate.regulation.matrices( reg.graph, 
                                           "./cache_reg_matrix.rda", FALSE )

reg.step <- reg.matrix$step
reg.flow <- reg.matrix$flow
path <- system.file("extdata", "sample_data.RData", package = "regflow")
load(path)

# Compute ERFA model
regul.matrices.refined <- erfa( reg.graph, reg.step,  stimulationTrain, train,
                                plot.dir=".", n.iteration=30,
                                stepsize=5, 
                                stimulationTest = stimulationTest,
                                expressionTest = test, fractionValid = 0.5)

# Compute MoAs for one drug perturbation
completeset <- cbind(train,test)
drug_to_test <- "mitoxantrone"
id <- which(colnames(completeset) == drug_to_test)

largestLFC=order(abs(completeset[,id]),decreasing = T)[1:10]
to.genes=rownames(completeset)[largestLFC]

alpha <- as.matrix(regul.matrices.refined$alpha[[id]])
rownames(alpha) <- rownames(reg.step)

from.genes<- rownames(stimulationTest)[stimulationTest[,drug_to_test]!=0]

erfa.discovery(reg.graph,regul.matrices.refined$step, regul.matrices.refined$flow,
               from.genes, to.genes,
               pathway.n.max = 10, pathway.gene.max = 1000, 
               plot.dir="./pathways", alpha=alpha, plot.moa = T)
 
}
