% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erfa.r
\name{erfa}
\alias{erfa}
\title{Estimation of ERFA model from the observed expression data}
\usage{
erfa(
  rg,
  regulation.step,
  stimulationTrain,
  expressionTrain,
  stepsize = 0.1,
  n.iteration = 20,
  plot.dir = NULL,
  plot.title = "Plot",
  stimulationTest = NULL,
  expressionTest = NULL,
  fractionValid = 0.5,
  regularisation = 0,
  plotIDtrain = NULL,
  plotIDtest = NULL,
  no_cores = 8
)
}
\arguments{
\item{rg}{An \link{igraph} graph object with the interaction graph, usually
obtained \code{\link[=build.regulation.graph]{build.regulation.graph()}}.}

\item{regulation.step}{Named numeric matrix with one-step regulation values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{stimulationTrain}{Named numeric matrix, defining stimulation values for one
or more genes over multiple experiments in the training data: genes in rows, experiments in column.
A value in the matrix should be equal to 0 if the gene is not stimulated, and non-zero
if the gene is stimulated.}

\item{expressionTrain}{Named numeric matrix, providing log-foldchanges of
expression values for training data:
genes in rows, experiments in columns.}

\item{stepsize}{Numeric value that specifies the step size in the
Gradient Descent algorithm}

\item{n.iteration}{Integer with the number of optimization iterations to be
run.}

\item{plot.dir}{Character string with the directory to save plots to. If
\code{NULL}, no plotting is carried out.}

\item{plot.title}{Character string with the title for all the plots.}

\item{stimulationTest}{Named numeric matrix, defining stimulation values for one
or more genes over multiple experiments in the test data: genes in rows, experiments in column.
A value in the matrix should be equal to 0 if the gene is not stimulated, and non-zero
if the gene is stimulated.}

\item{expressionTest}{Named numeric matrix, providing log-foldchanges of
expression values for test data:
genes in rows, experiments in columns.}

\item{fractionValid}{Should training data be split into training and validation?
The number specifies a proportion of split.}

\item{plotIDtrain}{index of perturbations in the training data
that shall be used
to plot dependence of log-foldchanges versus predicted regulatory flow.}

\item{plotIDtest}{index of perturbations in the test data
that shall be used
to plot dependence of log-foldchanges versus predicted regulatory flow.}

\item{no_cores}{Number of available threads for parallellization}
}
\value{
A list with two named numeric matrices:
\itemize{
\item \code{step}: Matrix with one-step regulation, refined.
\item \code{flow}: Matrix with all-step regulation, refined.
}
}
\description{
ERFA method uses Projected Gradient Descent algorithm with backtracking
to make predicted regulatory flow values closer to the observed log-fold changes.
The method returns updated step and flow (one-step and all-step) matrices.
Optimization is based on minimizing distance between the log fold changes and
the predicted regulatory flow.
}
\examples{
import.database.omnipath("./database_omnipath.rda", TRUE)
interacs <- get.interactions( "omnipath", 
                              "./database_omnipath.rda", "./table_interacs.csv",
                              "./cache_interacs.rda", FALSE )

reg.graph <- build.regulation.graph( interacs, NULL, "./cache_reg_graph.rda", 
                                     FALSE )

reg.matrix <- calculate.regulation.matrices( reg.graph, 
                                           "./cache_reg_matrix.rda", FALSE )

reg.step <- reg.matrix$step
reg.flow <- reg.matrix$flow
path <- system.file("extdata", "sample_data.RData", package = "regflow")
load(path)

# Compute ERFA model
regul.matrices.refined <- erfa( reg.graph, reg.step,  stimulationTrain, train,
                                plot.dir=".", n.iteration=30,
                                stepsize=5, 
                                stimulationTest = stimulationTest,
                                expressionTest = test, fractionValid = 0.5)

# Compute MoAs for one drug perturbation
completeset <- cbind(train,test)
drug_to_test <- "mitoxantrone"
id <- which(colnames(completeset) == drug_to_test)

largestLFC=order(abs(completeset[,id]),decreasing = T)[1:10]
to.genes=rownames(completeset)[largestLFC]

alpha <- as.matrix(regul.matrices.refined$alpha[[id]])
rownames(alpha) <- rownames(reg.step)

from.genes<- rownames(stimulationTest)[stimulationTest[,drug_to_test]!=0]

erfa.discovery(reg.graph,regul.matrices.refined$step, regul.matrices.refined$flow,
               from.genes, to.genes,
               pathway.n.max = 10, pathway.gene.max = 1000, 
               plot.dir="./pathways", alpha=alpha, plot.moa = T)
 
}
