% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_regulation_flow.r
\name{write.regulation.flow}
\alias{write.regulation.flow}
\title{Write table with regulation flow from or to a gene}
\usage{
write.regulation.flow(regulation.flow, gene, is.from, write.dir)
}
\arguments{
\item{regulation.flow}{A matrix with regulation flow values, obtained with
\code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{gene}{Character vector with gene names as source or target of
regulation.}

\item{is.from}{Boolean indicating:
\itemize{
\item \code{TRUE}: Downstream regulation from \code{gene}.
\item \code{FALSE}: Upstream regulation to \code{gene}.
}}

\item{write.dir}{Character string with the directory to write the table.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Writes a CSV table with the amount of regulation flow from or to a gene or
protein.
}
\details{
The call is ignored if \code{gene} is \code{NULL} or has no elements.
}
